Start over and refactor package structure to new design
