--- src/main/java/com/matzua/client/Constants.java
@@ -1,5 +1,0 @@
-package com.matzua.client;
-
-public interface Constants {
-    int ENTITY_PLAYER = 0;
-}
