--- src/main/java/com/matzua/client/Game.java
@@ -1,9 +1,0 @@
-package com.matzua.client;
-
-import lombok.Setter;
-import processing.core.PGraphics;
-
-public class Game {
-    @Setter
-    private PGraphics graphics;
-}
