+++ src/main/java/com/matzua/engine/core/EventManager.java
@@ -1,0 +1,9 @@
+package com.matzua.engine.core;
+
+import lombok.AllArgsConstructor;
+
+import javax.inject.Inject;
+
+@AllArgsConstructor(onConstructor_ = {@Inject})
+public class EventManager {
+}
