+++ src/main/java/com/matzua/game/example/Main.java
@@ -1,0 +1,9 @@
+package com.matzua.game.example;
+
+import com.matzua.game.example.dev.Dev;
+
+public class Main {
+    public static void main(String[] args) {
+        Dev.main(args);
+    }
+}
