+++ src/main/java/com/matzua/game/example/dagger/component/AppComponent.java
@@ -1,0 +1,10 @@
+package com.matzua.game.example.dagger.component;
+
+import com.matzua.engine.app.App;
+import com.matzua.game.example.dagger.module.CoreModule;
+import dagger.Component;
+
+@Component(modules = {CoreModule.class})
+public interface AppComponent {
+    App getApp();
+}
