+++ src/main/java/com/matzua/game/example/dagger/module/CoreModule.java
@@ -1,0 +1,7 @@
+package com.matzua.game.example.dagger.module;
+
+import dagger.Module;
+
+@Module
+public interface CoreModule {
+}
