--- src/main/java/com/matzua/util/MethodName.java
@@ -1,22 +1,0 @@
-package com.matzua.util;
-
-import static java.util.function.Function.identity;
-
-public interface MethodName {
-    static String at(int offset) {
-        return StackWalker.getInstance()
-                .walk(identity())
-                .skip(offset)
-                .findFirst()
-                .map(StackWalker.StackFrame::getMethodName)
-                .orElseThrow();
-    }
-
-    static String current() {
-        return at(2);
-    }
-
-    static String caller() {
-        return at(3);
-    }
-}
