+++ src/test/java/com/matzua/engine/app/AppTest.java
@@ -1,0 +1,4 @@
+package com.matzua.engine.app;
+
+public class AppTest {
+}
