+++ src/test/java/com/matzua/engine/core/EventManagerTest.java
@@ -1,0 +1,4 @@
+package com.matzua.engine.core;
+
+public class EventManagerTest {
+}
